Opportunistic refactor:
- Extract expression-evaluation code to its own file
- Test expression-evaluation using unit tests instead of integration-tests
