--- app/src/androidTest/java/com/hedvig/app/feature/embark/redirects/MultipleExpressionRedirectTest.kt
@@ -1,48 +1,0 @@
-package com.hedvig.app.feature.embark.redirects
-
-import com.hedvig.android.owldroid.graphql.EmbarkStoryQuery
-import com.hedvig.app.feature.embark.screens.EmbarkScreen
-import com.hedvig.app.feature.embark.ui.EmbarkActivity
-import com.hedvig.app.testdata.feature.embark.data.STORY_WITH_MULTIPLE_REDIRECTS
-import com.hedvig.app.util.ApolloCacheClearRule
-import com.hedvig.app.util.ApolloMockServerRule
-import com.hedvig.app.util.LazyActivityScenarioRule
-import com.hedvig.app.util.apolloResponse
-import com.hedvig.app.util.context
-import com.kaspersky.kaspresso.testcases.api.testcase.TestCase
-import io.github.kakaocup.kakao.screen.Screen.Companion.onScreen
-import org.junit.Rule
-import org.junit.Test
-
-class MultipleExpressionRedirectTest : TestCase() {
-    @get:Rule
-    val activityRule = LazyActivityScenarioRule(EmbarkActivity::class.java)
-
-    @get:Rule
-    val apolloMockServerRule = ApolloMockServerRule(
-        EmbarkStoryQuery.QUERY_DOCUMENT to apolloResponse { success(STORY_WITH_MULTIPLE_REDIRECTS) }
-    )
-
-    @get:Rule
-    val apolloCacheClearRule = ApolloCacheClearRule()
-
-    @Test
-    fun shouldRedirectOnPassageWithMultipleRedirect() = run {
-        activityRule.launch(
-            EmbarkActivity.newInstance(
-                context(),
-                this.javaClass.name,
-                "",
-            )
-        )
-
-        onScreen<EmbarkScreen> {
-            selectActions { firstChild<EmbarkScreen.SelectAction> { click() } }
-            messages {
-                firstChild<EmbarkScreen.MessageRow> {
-                    text { hasText("a third message") }
-                }
-            }
-        }
-    }
-}
