+++ app/src/main/java/com/hedvig/app/feature/embark/EmbarkViewModel.kt
@@ -9,13 +9,11 @@ import com.hedvig.android.owldroid.fragment.BasicExpressionFragment
 import com.hedvig.android.owldroid.fragment.ExpressionFragment
 import com.hedvig.android.owldroid.fragment.MessageFragment
 import com.hedvig.android.owldroid.graphql.EmbarkStoryQuery
-import com.hedvig.android.owldroid.type.EmbarkExpressionTypeBinary
-import com.hedvig.android.owldroid.type.EmbarkExpressionTypeMultiple
-import com.hedvig.android.owldroid.type.EmbarkExpressionTypeUnary
 import com.hedvig.android.owldroid.type.EmbarkExternalRedirectLocation
 import com.hedvig.app.authenticate.LoginStatus
 import com.hedvig.app.authenticate.LoginStatusService
 import com.hedvig.app.feature.embark.util.VariableExtractor
+import com.hedvig.app.feature.embark.util.evaluateExpression
 import com.hedvig.app.util.Percent
 import com.hedvig.app.util.getWithDotNotation
 import com.hedvig.app.util.plus
@@ -124,7 +122,7 @@ abstract class EmbarkViewModel(
             val nextPassage = story.passages.find { it.name == passageName }
             if (nextPassage?.redirects?.isNotEmpty() == true) {
                 nextPassage.redirects.forEach { redirect ->
-                    if (evaluateExpression(redirect.into()) is ExpressionResult.True) {
+                    if (evaluateExpression(redirect.into(), valueStore) is ExpressionResult.True) {
                         redirect.passedKeyValue?.let { (key, value) -> putInStore(key, value) }
                         redirect.to?.let { to ->
                             navigateToPassage(to)
@@ -426,7 +424,7 @@ abstract class EmbarkViewModel(
 
         val expressionText = message
             .expressions
-            .map { evaluateExpression(it.fragments.expressionFragment) }
+            .map { evaluateExpression(it.fragments.expressionFragment, valueStore) }
             .filterIsInstance<ExpressionResult.True>()
             .firstOrNull()
             ?.resultValue
@@ -435,110 +433,6 @@ abstract class EmbarkViewModel(
         return message.copy(
             text = interpolateMessage(expressionText)
         )
-    }
-
-    private fun evaluateExpression(expression: ExpressionFragment): ExpressionResult {
-        expression.fragments.basicExpressionFragment.asEmbarkExpressionUnary?.let { unaryExpression ->
-            return when (unaryExpression.unaryType) {
-                EmbarkExpressionTypeUnary.ALWAYS -> ExpressionResult.True(unaryExpression.text)
-                EmbarkExpressionTypeUnary.NEVER -> ExpressionResult.False
-                else -> ExpressionResult.False
-            }
-        }
-        expression.fragments.basicExpressionFragment.asEmbarkExpressionBinary?.let { binaryExpression ->
-            when (binaryExpression.binaryType) {
-                EmbarkExpressionTypeBinary.EQUALS -> {
-                    if (valueStore.get(binaryExpression.key) == binaryExpression.value) {
-                        return ExpressionResult.True(binaryExpression.text)
-                    }
-                }
-                EmbarkExpressionTypeBinary.NOT_EQUALS -> {
-                    val stored = valueStore.get(binaryExpression.key)
-                        ?: return ExpressionResult.False
-                    if (stored != binaryExpression.value) {
-                        return ExpressionResult.True(binaryExpression.text)
-                    }
-                }
-                EmbarkExpressionTypeBinary.MORE_THAN,
-                EmbarkExpressionTypeBinary.MORE_THAN_OR_EQUALS,
-                EmbarkExpressionTypeBinary.LESS_THAN,
-                EmbarkExpressionTypeBinary.LESS_THAN_OR_EQUALS,
-                -> {
-                    val storedAsInt = valueStore.get(binaryExpression.key)?.toIntOrNull()
-                        ?: return ExpressionResult.False
-                    val valueAsInt =
-                        binaryExpression.value.toIntOrNull() ?: return ExpressionResult.False
-
-                    val evaluatesToTrue = when (binaryExpression.binaryType) {
-                        EmbarkExpressionTypeBinary.MORE_THAN -> storedAsInt > valueAsInt
-                        EmbarkExpressionTypeBinary.MORE_THAN_OR_EQUALS -> storedAsInt >= valueAsInt
-                        EmbarkExpressionTypeBinary.LESS_THAN -> storedAsInt < valueAsInt
-                        EmbarkExpressionTypeBinary.LESS_THAN_OR_EQUALS -> storedAsInt <= valueAsInt
-                        else -> false
-                    }
-
-                    if (evaluatesToTrue) {
-                        return ExpressionResult.True(binaryExpression.text)
-                    }
-                }
-                else -> {
-                }
-            }
-            return ExpressionResult.False
-        }
-        expression.asEmbarkExpressionMultiple?.let { multipleExpression ->
-            val results =
-                multipleExpression.subExpressions.map {
-                    evaluateExpression(
-                        ExpressionFragment(
-                            fragments = ExpressionFragment.Fragments(it.fragments.basicExpressionFragment),
-                            asEmbarkExpressionMultiple = it.asEmbarkExpressionMultiple1?.let { asMulti ->
-                                ExpressionFragment.AsEmbarkExpressionMultiple(
-                                    multipleType = asMulti.multipleType,
-                                    text = asMulti.text,
-                                    subExpressions = asMulti.subExpressions.map { se ->
-                                        ExpressionFragment.SubExpression2(
-                                            fragments = ExpressionFragment.SubExpression2.Fragments(
-                                                se.fragments.basicExpressionFragment
-                                            ),
-                                            asEmbarkExpressionMultiple1 = se
-                                                .asEmbarkExpressionMultiple2?.let { asMulti2 ->
-                                                    ExpressionFragment.AsEmbarkExpressionMultiple1(
-                                                        multipleType = asMulti2.multipleType,
-                                                        text = asMulti2.text,
-                                                        subExpressions = asMulti2.subExpressions.map { se2 ->
-                                                            ExpressionFragment.SubExpression1(
-                                                                fragments = ExpressionFragment.SubExpression1.Fragments(
-                                                                    se2.fragments.basicExpressionFragment
-                                                                ),
-                                                                asEmbarkExpressionMultiple2 = null,
-                                                            )
-                                                        }
-                                                    )
-                                                }
-                                        )
-                                    }
-                                )
-                            },
-                        )
-                    )
-                }
-            when (multipleExpression.multipleType) {
-                EmbarkExpressionTypeMultiple.AND -> {
-                    if (results.all { it is ExpressionResult.True }) {
-                        return ExpressionResult.True(multipleExpression.text)
-                    }
-                }
-                EmbarkExpressionTypeMultiple.OR -> {
-                    if (results.any { it is ExpressionResult.True }) {
-                        return ExpressionResult.True(multipleExpression.text)
-                    }
-                }
-                else -> {
-                }
-            }
-        }
-        return ExpressionResult.False
     }
 
     private fun interpolateMessage(message: String, store: ValueStoreView = valueStore) =
